+++ src/main/java/com/timeflowsystem/security/config/ProjectSecurityConfig.java
@@ -25,9 +25,16 @@ public class ProjectSecurityConfig {
                 .authorizeHttpRequests(authorize -> authorize
                             .requestMatchers("/customers/**").permitAll()
                             .requestMatchers("/user-accounts/**").authenticated()
-                );
+                )
+                .formLogin(Customizer.withDefaults());
+
+        // This is a preferred global configuration:
+        //     http.exceptionHandling(ehc -> ehc.authenticationEntryPoint(new CustomBasicAuthenticationEntryPoint()));
+        // because it handles other ways of error 401 (Unauthorized).
+        // But, this exceptionHandling disables the .formLogin(Customizer.withDefaults()) configuration
 
         http.httpBasic(hbc -> hbc.authenticationEntryPoint(new CustomBasicAuthenticationEntryPoint()));
+
         return http.build();
     }